requirements.txt
streamlit
pandas
pytesseract
Pillow
pdfplumber
matplotlib
reportlab
spacy
thefuzz
python-Levenshtein
google-generativeai
sqlalchemy
passlib[bcrypt]
email-validator
python-dotenv

# MongoDB + Atlas support
pymongo
dnspython

# Optional: Better image handling
opencv-python

# Optional: To avoid warnings
numpy
